Queue size wrong calculation bugfix, camera framesize change, connection and battery sending, misc

+++ main/src/Services/Battery.h
@@ -105,7 +105,7 @@ public:
 
 	bool isShutdown() const;
 
-	uint8_t getPerc() const;
+	float getPerc() const;
 	Level getLevel() const;
 
 private:
